支持创建单独的类型
